Refactor Number Providers
