+++ src/test/kotlin/com/wesleyhome/test/jupiter/provider/datetime/InstantRangeSourceAbstractParameterDataProviderTest.kt
@@ -5,6 +5,7 @@ import com.wesleyhome.test.jupiter.annotations.InstantRangeSource
 import com.wesleyhome.test.jupiter.annotations.StringSource
 import com.wesleyhome.test.jupiter.provider.AnnotatedParameterDataProviderTest
 import com.wesleyhome.test.jupiter.provider.TestParameter
+import com.wesleyhome.test.jupiter.provider.number.step
 import com.wesleyhome.test.jupiter.provider.step
 import com.wesleyhome.test.jupiter.provider.temporalAmount
 import java.time.Instant
