+++ src/test/kotlin/com/wesleyhome/test/jupiter/provider/number/DoubleValueSourceDataProviderTest.kt
@@ -1,6 +1,8 @@
-package com.wesleyhome.test.jupiter.provider
+package com.wesleyhome.test.jupiter.provider.number
 
 import com.wesleyhome.test.jupiter.annotations.DoubleSource
+import com.wesleyhome.test.jupiter.provider.AnnotatedParameterDataProviderTest
+import com.wesleyhome.test.jupiter.provider.TestParameter
 import org.junit.jupiter.api.Test
 
 class DoubleValueSourceDataProviderTest :
